LayoutTests/fast/js/kde/evil-n-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (new Error()).message is ''
PASS ''.split(/.*/).length is 0
PASS window.successfullyParsed is undefined.

TEST COMPLETE
